requirements.txt
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==43.0.1
Deprecated==1.2.14
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
pycparser==2.21
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
smmap==5.0.1
typing_extensions==4.7.1
urllib3==2.0.7
wrapt==1.16.0
